# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: XAI Blockchain API
  version: 1.1.0
  description: 'Production REST API for the XAI blockchain node.


    All endpoints require authenticated transport in production. Use the summaries
    and schemas in this document to integrate wallets, explorers, trading engines,
    and operational tooling.'
  contact:
    name: XAI Support
    url: https://xai-blockchain.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: http://localhost:5000
  description: Local development
- url: https://testnet-api.xai-blockchain.io
  description: Testnet
- url: https://api.xai-blockchain.io
  description: Production
...
paths:
...
  /exchange/orders:
    get:
      tags:
      - Exchange
      summary: Order book snapshot
      description: Returns the top of book for both sides.
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  buy_orders:
                    type: array
                    items:
                      type: object
                  sell_orders:
                    type: array
                    items:
                      type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /exchange/place-order:
    post:
      tags:
      - Exchange
      summary: Place order
      description: Creates a limit order and optionally matches immediately.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeOrderInput'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  order:
                    type: object
                  matched:
                    type: array
                    items:
                      type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /exchange/cancel-order:
    post:
      tags:
      - Exchange
      summary: Cancel order
      description: Cancels an open order and unlocks funds.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeCancelInput'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  message:
                    type: string
                    example: Order cancelled successfully
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /exchange/my-orders/{address}:
    get:
      tags:
      - Exchange
      summary: List user orders
      description: Returns the order history for an address.
      parameters:
      - $ref: '#/components/parameters/AddressPathParam'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  orders:
                    type: array
                    items:
                      type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /exchange/trades:
    get:
      tags:
      - Exchange
      summary: Recent trades
      description: Returns recently executed trades.
      parameters:
      - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  trades:
                    type: array
                    items:
                      type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /exchange/deposit:
    post:
      tags:
      - Exchange
      summary: Deposit funds
      description: Credits an exchange wallet via manual or automated deposit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeTransferInput'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  result:
                    type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /exchange/withdraw:
    post:
      tags:
      - Exchange
      summary: Withdraw funds
      description: Initiates a withdrawal from the exchange wallet.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeTransferInput'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  result:
                    type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /exchange/balance/{address}:
    get:
      tags:
      - Exchange
      summary: All balances
      description: Returns all currency balances (available and locked) for an address.
      parameters:
      - $ref: '#/components/parameters/AddressPathParam'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  balances:
                    type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
...
components:
  parameters:
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
      description: Maximum number of records to return.
...
    AddressPathParam:
      name: address
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
        code:
          type: string
...
    ExchangeOrderInput:
      type: object
      required:
      - address
      - order_type
      - pair
      - price
      - amount
      properties:
        address:
          type: string
        order_type:
          type: string
        pair:
          type: string
        price:
          type: number
        amount:
          type: number
    ExchangeCancelInput:
      type: object
      required:
      - order_id
      properties:
        order_id:
          type: string
...
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: X-API-Key
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- api_key: []
- bearer_auth: []